requirements.txt
python>=3.10
xgboost>=2.0
shap>=0.43
dice-ml>=0.11
pandas>=2.0
scikit-learn>=1.3
streamlit>=1.30
matplotlib>=3.8
seaborn>=0.13
pyyaml
joblib
numpy>=1.24
scipy>=1.11
pytest>=7.0
